Keep gzip terminal output in its own helper.
The codec does not belong in the coalesce-window parser, and the gzip ack is a JSON control message like close reasons.
